Refactor dtype of total_party_max_hit_points to np.float32 for consistency

+++ pokegym/environment.py
@@ -276,7 +276,7 @@ class Environment(Base):
             "each_pokemon_health_points": spaces.Box(low = 0, high = 99, shape=(6,), dtype=np.uint8),
             "each_pokemon_max_health_points": spaces.Box(low = 0, high = 99, shape=(6,), dtype=np.uint8),
             "total_party_health_points": spaces.Box(low = 0, high = 99, shape=(1,), dtype=np.uint8),
-            "total_party_max_hit_points": spaces.Box(low = 0, high = 1, shape=(1,), dtype=np.float16),
+            "total_party_max_hit_points": spaces.Box(low = 0, high = 1, shape=(1,), dtype=np.float32),
             "low_health_alarm": spaces.Box(low = 0, high = 1, shape=(1,), dtype=np.uint8),
             "opponent_pokemon_levels": spaces.Box(low = 0, high = 100, shape=(6,), dtype=np.uint8),
             "total_number_of_items": spaces.Box(low = 0, high = 64, shape=(1,), dtype=np.uint8),
